01_insert: var j negative check

+++ 01_sort/01_insert/sort.cpp
@@ -6,7 +6,7 @@ void sort_insert(std::vector<int> &arr, bool (*cmp_func)(int, int))
     for (size_t i = 0; i < arr.size(); i++)
     {
         int tmp = arr[i];
-        size_t j = i - 1;
+        int j = i - 1;
         while (j >= 0 && cmp_func(tmp, arr[j]))
         {
             arr[j + 1] = arr[j];
